refactored track hours to be track volunteers for all files and variables.
