+++ app/controllers/events/routes.py
@@ -4,7 +4,6 @@ from flask import Flask, redirect, flash
 from app.controllers.events import events_bp
 from app.logic.events import getEvents
 from app.logic.getUpcomingEvents import getUpcomingEventsForUser
-from app.logic.updateTrackHours import updateTrackHours
 
 
 @events_bp.route('/events', methods=['GET'])
