+++ app/static/js/trackVolunteers.js
@@ -1,7 +1,7 @@
 $(document).ready(function(){
 
 // Search functionalities from the volunteer table in the UI
-  $("#trackHoursInput").on("keyup", function() {
+  $("#trackVolunteersInput").on("keyup", function() {
     var value = $(this).val().toLowerCase();
     $("#volunteerTable tr").filter(function() {
       $(this).toggle($(this).text().toLowerCase().indexOf(value) > -1)
@@ -54,7 +54,7 @@ function removeVolunteerFromEvent (deleteIcon){
 }
 
 // Search functionalities from the user table in the database
-function searchTrackHoursVolunteers(){
+function searchVolunteers(){
 
   $("#selectVolunteerButton").prop('disabled', true)
   var query = $("#addVolunteerInput").val()
@@ -64,7 +64,7 @@ function searchTrackHoursVolunteers(){
     minLength: 2,
     source: function(request, response){
       $.ajax({
-        url: "/searchTrackHoursVolunteers/" + query,
+        url: "/searchVolunteers/" + query,
         type: "GET",
         dataType: "json",
         success: function(dictToJSON) {
@@ -86,7 +86,7 @@ function searchTrackHoursVolunteers(){
   });
 };
 
-function toggleTrackHoursInputBox(checkbox){
+function toggleVolunteersInputBox(checkbox){
   username =  checkbox.id.substring(9) //get everything after the 9th character
   inputFieldID = 'inputHours_'+username
